Catch and Throw ClassCastException to provide details of the column with the issue
Without catching and throwing this exception it is handled by an catch
much higher in the call stack which sees it as an 'unknown' exception
and has no details of the column with the issue.
This fixes that to provide more information in the log.
Appears to log mutliple times (maybe just my logging) - should not be
because it's in a for loop as the exception being raised should break
out of the for loop