Fix for a problem with sqlserver and datanucleus-rdbms-useDefaultSqlType
If we are working with a ms sqlserver and switch the datanucleus option
"datanucleus.rdbms.useDefaultSqlType" to `false` the varbinary problem
returns, because the bad data fetched from the ms jdbc driver is back
again.

So it is the best to remove the varbinary definition provided by the
driver and use the defined one in the sqlserveradpter introduced
earlier.

The definition is slightly modified.