Remove column mapping to JDBC type FLOAT for SQL Server
The JDBC driver does not provide this type. According to https://learn.microsoft.com/en-us/sql/connect/jdbc/using-basic-data-types?view=sql-server-ver16, the SQL type `FLOAT` is mapped to the JDBC type `DOUBLE`.

dbinfo output of the schema tool is provided for the last three major versions of the JDBC driver.

Signed-off-by: nscuro <nscuro@protonmail.com>